requirements.txt
awsebcli==3.12.4
blessed==1.14.2
boto3==1.6.3
botocore==1.9.12
cement==2.8.2
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.2
django-cors-headers==2.2.0
django-filter==1.1.0
django-rest-swagger==2.1.2
django-storages==1.6.5
djangorestframework==3.7.7
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
pathspec==0.5.5
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
requests==2.9.1
s3transfer==0.1.13
semantic-version==2.5.0
simplejson==3.13.2
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.47.0
